requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
